Create README.md
R-Markdown document that compares code in R and Python for the most common and basic functions in exploratory data analysis. The code in the R-Markdown document is 100% reproducible, so you can run it yourself to see how it works.
The dataset I used is a real-world dataset from Public Health Scotland that contains information on community pharmacy activity. After running the code, you will find the top 40 prescribed items in Scotland in February 2023.
So, whether you're a Python user who needs to use R, or an R user who wants to learn Python, I encourage you to check out my RMarkdown document. It's a great way to learn how to use both languages for data analysis. The code is 100% reproducible, so you can run it yourself to see how it works.